Reducing theme specificity
Try this revised code for a ffew days before applying it to see if it cocks anything up. There are some things in this code that is exclusively for  dark theme when honestly, it _doesn't even matter_ and  would work for both.

This code is expressly so I don't have to work around existing changes here to make future stuff happen.